+++ lib/src/constants/constants.dart
@@ -1,10 +1,18 @@
-// enum ToastDirection {
-//   ltr,
-//   rtl,
-// }
-//
-// enum ToastPosition {
-//   top,
-//   center,
-//   bottom,
-// }
+/// Position of the toast on screen.
+enum ToastPosition {
+  topLeft,
+  topCenter,
+  topRight,
+  centerLeft,
+  center,
+  centerRight,
+  bottomLeft,
+  bottomCenter,
+  bottomRight,
+}
+
+/// Text direction for the toast content.
+enum ToastDirection {
+  ltr,
+  rtl,
+}